kfs: arch: get rid of u32s when not needed.
Most of the code in order to be arch independent
it cannot use u32, i32 all over the place so lets
convert it to usize where necessary to go one step
closer to supporting another arch.

+++ src/drivers/bdev.zig
@@ -30,14 +30,14 @@ fn bdev_close(base: *krn.fs.File) void {
     _ = base;
 }
 
-fn bdev_write(base: *krn.fs.File, buf: [*]const u8, size: u32) !u32 {
+fn bdev_write(base: *krn.fs.File, buf: [*]const u8, size: usize) !usize {
     _ = base;
     _ = buf;
     _ = size;
     return 0;
 }
 
-fn bdev_read(base: *krn.fs.File, buf: [*]u8, size: u32) !u32 {
+fn bdev_read(base: *krn.fs.File, buf: [*]u8, size: usize) !usize {
     _ = base;
     _ = buf;
     _ = size;
