+++ src/module/api.zig
@@ -10,12 +10,12 @@ pub fn init() void {
 pub export const error_offset: u16 = @intFromError(krn.errors.PosixError.EPERM);
 
 // Memory
-pub export fn kheap_alloc(size: u32, contig: bool, user: bool) u32 {
+pub export fn kheap_alloc(size: usize, contig: bool, user: bool) usize {
     const addr = krn.mm.kheap.alloc(size, contig, user) catch return 0;
     return addr;
 }
 
-pub export fn kheap_free(addr: u32) void {
+pub export fn kheap_free(addr: usize) void {
     krn.mm.kheap.free(addr);
 }
